+++ src/components/FormikForm/FormikForm.tsx
@@ -1,4 +1,4 @@
-import React, { useCallback, useMemo } from 'react'
+import React, { ReactNode, useCallback, useContext, useMemo } from 'react'
 import { connect, FormikContext, Form as FrmForm, Formik as FrmFormik, FormikConfig, FormikActions } from 'formik'
 import { SelectOption, Select as UiKitSelect, SelectProps as UiKitSelectProps } from '../Select/Select'
 import {
@@ -56,6 +56,7 @@ import {
 } from '../ExpressionInput/ExpressionInput'
 import { DataTooltipInterface } from '../../frameworks/Tooltip/types'
 import { HarnessDocTooltip } from '../../frameworks/Tooltip/Tooltip'
+import { FormikTooltipContext } from './FormikTooltipContext'
 
 const IsOptionLabel = '(optional)'
 const isObject = (obj: any): boolean => obj !== null && typeof obj === 'object'
@@ -71,6 +72,7 @@ export const getDefaultAutoCompleteValue = (): string => 'off'
 export interface FormikExtended<T> extends FormikContext<T> {
   disabled?: boolean
   inline?: boolean
+  formName?: string
 }
 
 export interface FormikContextProps<T> {
@@ -78,7 +80,24 @@ export interface FormikContextProps<T> {
   optionalLabel?: string
   isOptional?: boolean // default to false
   autoComplete?: string
-  tooltipProps?: DataTooltipInterface
+  tooltipProps?: DataTooltipInterface // todo mark it as mandatory
+}
+
+export const getFormFieldLabel = (
+  label: ReactNode | string | undefined,
+  fieldName: string,
+  props: FormikContextProps<any>,
+  css?: string
+): ReactNode | string | undefined => {
+  const optionalLabel = props.optionalLabel || IsOptionLabel
+  const labelText = !props.isOptional ? label : `${label} ${optionalLabel}`
+  if (!labelText) {
+    return labelText
+  }
+  const tooltipContext = useContext(FormikTooltipContext)
+  const dataTooltipId =
+    props.tooltipProps?.dataTooltipId || (tooltipContext?.formName ? `${tooltipContext?.formName}_${fieldName}` : '')
+  return <HarnessDocTooltip tooltipId={dataTooltipId} labelText={labelText} className={css || ''} />
 }
 
 export interface TagInputProps<T> extends Omit<IFormGroupProps, 'labelFor' | 'items'> {
@@ -91,7 +110,7 @@ export interface TagInputProps<T> extends Omit<IFormGroupProps, 'labelFor' | 'it
 }
 
 function TagInput<T>(props: TagInputProps<T> & FormikContextProps<any>) {
-  const { formik, name, isOptional = false, ...restProps } = props
+  const { formik, name, ...restProps } = props
   const hasError = errorCheck(name, formik)
   const {
     intent = hasError ? Intent.DANGER : Intent.NONE,
@@ -99,23 +118,18 @@ function TagInput<T>(props: TagInputProps<T> & FormikContextProps<any>) {
     disabled = formik?.disabled,
     items,
     label,
-    optionalLabel = IsOptionLabel,
     labelFor,
     itemFromNewTag,
     inline = formik?.inline,
     tagInputProps,
     onChange,
-    tooltipProps,
     ...rest
   } = restProps
 
-  const labelText = !isOptional ? label : `${label} ${optionalLabel}`
   return (
     <FormGroup
       labelFor={name}
-      label={
-        labelText ? <HarnessDocTooltip tooltipId={tooltipProps?.dataTooltipId} labelText={labelText} /> : labelText
-      }
+      label={getFormFieldLabel(label, name, props)}
       helperText={helperText}
       intent={intent}
       disabled={disabled}
@@ -156,16 +170,7 @@ const MENTIONS_DEFAULT: MentionsInfo = {
 }
 
 function KVTagInput(props: KVTagInputProps & FormikContextProps<any>) {
-  const {
-    formik,
-    name,
-    mentionsInfo,
-    tagsProps,
-    isOptional = false,
-    optionalLabel = IsOptionLabel,
-    tooltipProps,
-    ...restProps
-  } = props
+  const { formik, name, mentionsInfo, tagsProps, ...restProps } = props
   const hasError = errorCheck(name, formik)
   const {
     intent = hasError ? Intent.DANGER : Intent.NONE,
@@ -182,14 +187,10 @@ function KVTagInput(props: KVTagInputProps & FormikContextProps<any>) {
     return () => unregister(mentionsType)
   }, [])
 
-  const labelText = !isOptional ? label : `${label} ${optionalLabel}`
-
   return (
     <FormGroup
       labelFor={name}
-      label={
-        labelText ? <HarnessDocTooltip tooltipId={tooltipProps?.dataTooltipId} labelText={labelText} /> : labelText
-      }
+      label={getFormFieldLabel(label, name, props)}
       helperText={helperText}
       intent={intent}
       disabled={disabled}
@@ -234,16 +235,7 @@ export interface MultiInputProps extends Omit<IFormGroupProps, 'labelFor' | 'ite
 }
 
 function MultiInput(props: MultiInputProps & FormikContextProps<any>) {
-  const {
-    formik,
-    name,
-    mentionsInfo,
-    tagsProps,
-    optionalLabel = IsOptionLabel,
-    isOptional = false,
-    tooltipProps,
-    ...restProps
-  } = props
+  const { formik, name, mentionsInfo, tagsProps, ...restProps } = props
   const hasError = errorCheck(name, formik)
   const {
     intent = hasError ? Intent.DANGER : Intent.NONE,
@@ -260,13 +252,9 @@ function MultiInput(props: MultiInputProps & FormikContextProps<any>) {
     return () => unregister(mentionsType)
   }, [])
 
-  const labelText = !isOptional ? label : `${label} ${optionalLabel}`
-
   return (
     <FormGroup
-      label={
-        labelText ? <HarnessDocTooltip tooltipId={tooltipProps?.dataTooltipId} labelText={labelText} /> : labelText
-      }
+      label={getFormFieldLabel(label, name, props)}
       labelFor={name}
       helperText={helperText}
       intent={intent}
@@ -300,7 +288,7 @@ export interface CustomRenderProps extends Omit<IFormGroupProps, 'labelFor'> {
 }
 
 const CustomRender = (props: CustomRenderProps & FormikContextProps<any>) => {
-  const { formik, name, isOptional = false, optionalLabel = IsOptionLabel, ...restProps } = props
+  const { formik, name, ...restProps } = props
   const hasError = errorCheck(name, formik)
   const {
     intent = hasError ? Intent.DANGER : Intent.NONE,
@@ -309,16 +297,12 @@ const CustomRender = (props: CustomRenderProps & FormikContextProps<any>) => {
     inline = formik?.inline,
     label,
     render,
-    tooltipProps,
     ...rest
   } = restProps
 
-  const labelText = !isOptional ? label : `${label} ${optionalLabel}`
   return (
     <FormGroup
-      label={
-        labelText ? <HarnessDocTooltip tooltipId={tooltipProps?.dataTooltipId} labelText={labelText} /> : labelText
-      }
+      label={getFormFieldLabel(label, name, props)}
       labelFor={name}
       helperText={helperText}
       intent={intent}
@@ -342,7 +326,7 @@ export interface FileInputProps extends Omit<IFormGroupProps, 'labelFor'> {
 }
 
 const FileInput = (props: FileInputProps & FormikContextProps<any>) => {
-  const { formik, name, isOptional = false, optionalLabel = IsOptionLabel, ...restProps } = props
+  const { formik, name, ...restProps } = props
   const hasError = errorCheck(name, formik)
   const {
     intent = hasError ? Intent.DANGER : Intent.NONE,
@@ -356,16 +340,12 @@ const FileInput = (props: FileInputProps & FormikContextProps<any>) => {
     onChange,
     inputProps,
     multiple = false,
-    tooltipProps,
     ...rest
   } = restProps
 
-  const labelText = !isOptional ? label : `${label} ${optionalLabel}`
   return (
     <FormGroup
-      label={
-        labelText ? <HarnessDocTooltip tooltipId={tooltipProps?.dataTooltipId} labelText={labelText} /> : labelText
-      }
+      label={getFormFieldLabel(label, name, props)}
       labelFor={name}
       helperText={helperText}
       intent={intent}
@@ -411,12 +391,9 @@ const RadioGroup = (props: RadioGroupProps & FormikContextProps<any>) => {
     disabled = formik?.disabled,
     inline = formik?.inline,
     items = [],
-    isOptional = false,
-    optionalLabel = IsOptionLabel,
     label,
     radioGroup,
     onChange,
-    tooltipProps,
     ...rest
   } = props
 
@@ -427,12 +404,9 @@ const RadioGroup = (props: RadioGroupProps & FormikContextProps<any>) => {
       className: cx(cssRadio.radio, className)
     }
   })
-  const labelText = !isOptional ? label : `${label} ${optionalLabel}`
   return (
     <FormGroup
-      label={
-        labelText ? <HarnessDocTooltip tooltipId={tooltipProps?.dataTooltipId} labelText={labelText} /> : labelText
-      }
+      label={getFormFieldLabel(label, name, props)}
       labelFor={name}
       helperText={helperText}
       intent={intent}
@@ -460,7 +434,7 @@ export interface CheckboxProps extends UiKitCheckboxProps, Omit<IFormGroupProps,
 }
 
 const CheckBox = (props: CheckboxProps & FormikContextProps<any>) => {
-  const { formik, name, label, isOptional, optionalLabel, ...restProps } = props
+  const { formik, name, label, ...restProps } = props
   const hasError = errorCheck(name, formik)
   const {
     intent = hasError ? Intent.DANGER : Intent.NONE,
@@ -469,10 +443,8 @@ const CheckBox = (props: CheckboxProps & FormikContextProps<any>) => {
     inline = formik?.inline,
     onChange,
     className = '',
-    tooltipProps,
     ...rest
   } = restProps
-  const labelText = !isOptional ? label : `${label} ${optionalLabel}`
   return (
     <FormGroup labelFor={name} helperText={helperText} intent={intent} disabled={disabled} {...rest}>
       <UiKitCheckbox
@@ -481,17 +453,7 @@ const CheckBox = (props: CheckboxProps & FormikContextProps<any>) => {
         name={name}
         // eslint-disable-next-line
         // @ts-ignore
-        label={
-          labelText ? (
-            <HarnessDocTooltip
-              className={css.checkBoxDocTooltipLabel}
-              tooltipId={tooltipProps?.dataTooltipId}
-              labelText={labelText}
-            />
-          ) : (
-            labelText
-          )
-        }
+        label={getFormFieldLabel(label, name, props, css.checkBoxDocTooltipLabel)}
         inline={inline}
         disabled={disabled}
         checked={get(formik?.values, name)}
@@ -514,7 +476,7 @@ export interface MultiSelectProps extends Omit<IFormGroupProps, 'labelFor'> {
 }
 
 const MultiSelect = (props: MultiSelectProps & FormikContextProps<any>) => {
-  const { formik, name, isOptional = false, optionalLabel = IsOptionLabel, ...restProps } = props
+  const { formik, name, ...restProps } = props
   const hasError = errorCheck(name, formik)
   const {
     intent = hasError ? Intent.DANGER : Intent.NONE,
@@ -527,18 +489,14 @@ const MultiSelect = (props: MultiSelectProps & FormikContextProps<any>) => {
     placeholder,
     multiSelectProps,
     onChange,
-    tooltipProps,
     ...rest
   } = restProps
 
   const formikValue = get(formik?.values, name, [])
   const autoComplete = props.autoComplete || getDefaultAutoCompleteValue()
-  const labelText = !isOptional ? label : `${label} ${optionalLabel}`
   return (
     <FormGroup
-      label={
-        labelText ? <HarnessDocTooltip tooltipId={tooltipProps?.dataTooltipId} labelText={labelText} /> : labelText
-      }
+      label={getFormFieldLabel(label, name, props)}
       labelFor={name}
       helperText={helperText}
       intent={intent}
@@ -583,7 +541,7 @@ export interface SelectProps extends Omit<IFormGroupProps, 'labelFor'> {
 }
 
 const Select = (props: SelectProps & FormikContextProps<any>) => {
-  const { formik, name, isOptional = false, optionalLabel = IsOptionLabel, ...restProps } = props
+  const { formik, name, ...restProps } = props
   const hasError = errorCheck(name, formik)
   const {
     intent = hasError ? Intent.DANGER : Intent.NONE,
@@ -597,17 +555,13 @@ const Select = (props: SelectProps & FormikContextProps<any>) => {
     selectProps,
     onChange,
     value,
-    tooltipProps,
     ...rest
   } = restProps
 
   const autoComplete = props.autoComplete || getDefaultAutoCompleteValue()
-  const labelText = !isOptional ? label : `${label} ${optionalLabel}`
   return (
     <FormGroup
-      label={
-        labelText ? <HarnessDocTooltip tooltipId={tooltipProps?.dataTooltipId} labelText={labelText} /> : labelText
-      }
+      label={getFormFieldLabel(label, name, props)}
       labelFor={name}
       helperText={helperText}
       intent={intent}
@@ -646,7 +600,7 @@ export interface TextProps extends Omit<IFormGroupProps, 'labelFor'> {
 }
 
 const Text = (props: TextProps & FormikContextProps<any>) => {
-  const { formik, name, isOptional = false, optionalLabel = IsOptionLabel, ...restProps } = props
+  const { formik, name, ...restProps } = props
   const hasError = errorCheck(name, formik)
   const {
     intent = hasError ? Intent.DANGER : Intent.NONE,
@@ -657,16 +611,12 @@ const Text = (props: TextProps & FormikContextProps<any>) => {
     label,
     placeholder,
     onChange,
-    tooltipProps,
     ...rest
   } = restProps
   const autoComplete = props.autoComplete || getDefaultAutoCompleteValue()
-  const labelText = !isOptional ? label : `${label} ${optionalLabel}`
   return (
     <FormGroup
-      label={
-        labelText ? <HarnessDocTooltip tooltipId={tooltipProps?.dataTooltipId} labelText={labelText} /> : labelText
-      }
+      label={getFormFieldLabel(label, name, props)}
       labelFor={name}
       helperText={helperText}
       intent={intent}
@@ -707,18 +657,7 @@ export interface ExpressionInputProps extends Omit<IFormGroupProps, 'labelFor'> 
 }
 
 const ExpressionInput = (props: ExpressionInputProps & FormikContextProps<any>) => {
-  const {
-    formik,
-    name,
-    items = [],
-    placeholder,
-    expressionInputProps,
-    onChange,
-    isOptional,
-    optionalLabel = IsOptionLabel,
-    tooltipProps,
-    ...restProps
-  } = props
+  const { formik, name, items = [], placeholder, expressionInputProps, onChange, ...restProps } = props
   const hasError = errorCheck(name, formik)
   const {
     intent = hasError ? Intent.DANGER : Intent.NONE,
@@ -730,14 +669,10 @@ const ExpressionInput = (props: ExpressionInputProps & FormikContextProps<any>) 
   } = restProps
 
   const autoComplete = props.autoComplete || getDefaultAutoCompleteValue()
-  const labelText = !isOptional ? label : `${label} ${optionalLabel}`
-
   return (
     <FormGroup
       labelFor={name}
-      label={
-        labelText ? <HarnessDocTooltip tooltipId={tooltipProps?.dataTooltipId} labelText={labelText} /> : labelText
-      }
+      label={getFormFieldLabel(label, name, props)}
       helperText={helperText}
       intent={intent}
       disabled={disabled}
@@ -767,7 +702,7 @@ export interface TextAreaProps extends Omit<IFormGroupProps, 'labelFor'> {
 }
 
 const TextArea = (props: TextAreaProps & FormikContextProps<any>) => {
-  const { formik, name, autoFocus, isOptional = false, optionalLabel = IsOptionLabel, ...restProps } = props
+  const { formik, name, autoFocus, ...restProps } = props
   const hasError = errorCheck(name, formik)
   const {
     intent = hasError ? Intent.DANGER : Intent.NONE,
@@ -778,17 +713,13 @@ const TextArea = (props: TextAreaProps & FormikContextProps<any>) => {
     label,
     textArea,
     onChange,
-    tooltipProps,
     ...rest
   } = restProps
 
   const autoComplete = props.autoComplete || getDefaultAutoCompleteValue()
-  const labelText = !isOptional ? label : `${label} ${optionalLabel}`
   return (
     <FormGroup
-      label={
-        labelText ? <HarnessDocTooltip tooltipId={tooltipProps?.dataTooltipId} labelText={labelText} /> : labelText
-      }
+      label={getFormFieldLabel(label, name, props)}
       labelFor={name}
       helperText={helperText}
       intent={intent}
@@ -851,6 +782,7 @@ export interface FormikProps<Values> extends Omit<FormikConfig<Values>, 'onSubmi
   formLoading?: true
   render?: (props: FormikExtended<Values>) => React.ReactNode
   onSubmit: (values: Values, formikActions: FormikActions<Values>) => void | Promise<Values>
+  formName?: string
 }
 
 export function Formik<Values = Record<string, unknown>>(props: FormikProps<Values>): React.ReactElement {
@@ -901,9 +833,12 @@ export function Formik<Values = Record<string, unknown>>(props: FormikProps<Valu
       render: functionRenderLocal
     }
   }
+
   return (
     <FrmFormik {...rest} {...renderProps} onSubmit={onSubmitLocal}>
-      {!render && !isFunction(children) && <OverlaySpinner show={isFormLoading}>{children}</OverlaySpinner>}
+      <FormikTooltipContext.Provider value={{ formName: props.formName }}>
+        {!render && !isFunction(children) && <OverlaySpinner show={isFormLoading}>{children}</OverlaySpinner>}
+      </FormikTooltipContext.Provider>
     </FrmFormik>
   )
 }
@@ -914,7 +849,7 @@ export interface FormColorPickerProps extends ColorPickerProps, Omit<IFormGroupP
 }
 
 const FormColorPicker = (props: FormColorPickerProps & FormikContextProps<any>) => {
-  const { formik, name, isOptional = false, optionalLabel = IsOptionLabel, ...restProps } = props
+  const { formik, name, ...restProps } = props
   const hasError = errorCheck(name, formik)
   const {
     intent = hasError ? Intent.DANGER : Intent.NONE,
@@ -926,7 +861,7 @@ const FormColorPicker = (props: FormColorPickerProps & FormikContextProps<any>) 
   } = restProps
   return (
     <FormGroup
-      label={!isOptional ? label : `${label} ${optionalLabel}`}
+      label={getFormFieldLabel(label, name, props)}
       labelFor={name}
       helperText={helperText}
       intent={intent}
@@ -955,23 +890,13 @@ export interface FormMultiTypeInputProps extends Omit<IFormGroupProps, 'labelFor
 }
 
 const FormMultiTypeInput = (props: FormMultiTypeInputProps & FormikContextProps<any>) => {
-  const {
-    formik,
-    name,
-    selectItems,
-    placeholder,
-    multiTypeInputProps,
-    isOptional = false,
-    optionalLabel = IsOptionLabel,
-    ...restProps
-  } = props
+  const { formik, name, selectItems, placeholder, multiTypeInputProps, ...restProps } = props
   const hasError = errorCheck(name, formik)
   const {
     intent = hasError ? Intent.DANGER : Intent.NONE,
     helperText = hasError ? get(formik?.errors, name) : null,
     disabled = formik?.disabled,
     label,
-    tooltipProps,
     ...rest
   } = restProps
   const onChangeCallback: MultiTypeInputProps['onChange'] = useCallback(
@@ -983,12 +908,9 @@ const FormMultiTypeInput = (props: FormMultiTypeInputProps & FormikContextProps<
     [formik, multiTypeInputProps]
   )
   const autoComplete = props.autoComplete || getDefaultAutoCompleteValue()
-  const labelText = !isOptional ? label : `${label} ${optionalLabel}`
   return (
     <FormGroup
-      label={
-        labelText ? <HarnessDocTooltip tooltipId={tooltipProps?.dataTooltipId} labelText={labelText} /> : labelText
-      }
+      label={getFormFieldLabel(label, name, props)}
       labelFor={name}
       helperText={helperText}
       intent={intent}
@@ -1027,17 +949,7 @@ export interface FormMultiSelectTypeInputProps extends Omit<IFormGroupProps, 'la
 }
 
 const FormMultiSelectTypeInput = (props: FormMultiSelectTypeInputProps & FormikContextProps<any>) => {
-  const {
-    formik,
-    name,
-    selectItems,
-    placeholder,
-    multiSelectTypeInputProps,
-    isOptional = false,
-    optionalLabel = IsOptionLabel,
-    tooltipProps,
-    ...restProps
-  } = props
+  const { formik, name, selectItems, placeholder, multiSelectTypeInputProps, ...restProps } = props
   const hasError = errorCheck(name, formik)
   const {
     intent = hasError ? Intent.DANGER : Intent.NONE,
@@ -1047,12 +959,9 @@ const FormMultiSelectTypeInput = (props: FormMultiSelectTypeInputProps & FormikC
     ...rest
   } = restProps
   const autoComplete = props.autoComplete || getDefaultAutoCompleteValue()
-  const labelText = !isOptional ? label : `${label} ${optionalLabel}`
   return (
     <FormGroup
-      label={
-        labelText ? <HarnessDocTooltip tooltipId={tooltipProps?.dataTooltipId} labelText={labelText} /> : labelText
-      }
+      label={getFormFieldLabel(label, name, props)}
       labelFor={name}
       helperText={helperText}
       intent={intent}
@@ -1097,23 +1006,13 @@ export interface FormMultiTextTypeInputProps extends Omit<IFormGroupProps, 'labe
 }
 
 const FormMultiTextTypeInput = (props: FormMultiTextTypeInputProps & FormikContextProps<any>) => {
-  const {
-    formik,
-    name,
-    placeholder,
-    multiTextInputProps,
-    onChange,
-    isOptional = false,
-    optionalLabel = IsOptionLabel,
-    ...restProps
-  } = props
+  const { formik, name, placeholder, multiTextInputProps, onChange, ...restProps } = props
   const hasError = errorCheck(name, formik)
   const {
     intent = hasError ? Intent.DANGER : Intent.NONE,
     helperText = hasError ? get(formik?.errors, name) : null,
     disabled = formik?.disabled,
     label,
-    tooltipProps,
     ...rest
   } = restProps
   const value = get(formik?.values, name, '')
@@ -1129,12 +1028,9 @@ const FormMultiTextTypeInput = (props: FormMultiTextTypeInputProps & FormikConte
   )
 
   const autoComplete = props.autoComplete || getDefaultAutoCompleteValue()
-  const labelText = !isOptional ? label : `${label} ${optionalLabel}`
   return (
     <FormGroup
-      label={
-        labelText ? <HarnessDocTooltip tooltipId={tooltipProps?.dataTooltipId} labelText={labelText} /> : labelText
-      }
+      label={getFormFieldLabel(label, name, props)}
       labelFor={name}
       helperText={helperText}
       intent={intent}
@@ -1165,7 +1061,7 @@ export interface FormCategorizedSelect extends Omit<IFormGroupProps, 'labelFor'>
 }
 
 const FormCategorizedSelect = (props: FormCategorizedSelect & FormikContextProps<any>) => {
-  const { formik, name, isOptional = false, optionalLabel = IsOptionLabel, ...restProps } = props
+  const { formik, name, ...restProps } = props
   const hasError = errorCheck(name, formik)
   const {
     intent = hasError ? Intent.DANGER : Intent.NONE,
@@ -1177,17 +1073,13 @@ const FormCategorizedSelect = (props: FormCategorizedSelect & FormikContextProps
     inline = formik?.inline,
     categorizedSelectProps,
     onChange,
-    tooltipProps,
     ...rest
   } = restProps
 
   const value = get(formik?.values, name)
-  const labelText = !isOptional ? label : `${label} ${optionalLabel}`
   return (
     <FormGroup
-      label={
-        labelText ? <HarnessDocTooltip tooltipId={tooltipProps?.dataTooltipId} labelText={labelText} /> : labelText
-      }
+      label={getFormFieldLabel(label, name, props)}
       labelFor={name}
       helperText={helperText}
       intent={intent}
@@ -1228,7 +1120,7 @@ export interface FormSelectWithSubviewProps extends Omit<IFormGroupProps, 'label
 }
 
 const FormSelectWithSubview = (props: FormSelectWithSubviewProps & FormikContextProps<any>) => {
-  const { formik, name, isOptional = false, optionalLabel = IsOptionLabel, ...restProps } = props
+  const { formik, name, ...restProps } = props
   const hasError = errorCheck(name, formik)
   const {
     intent = hasError ? Intent.DANGER : Intent.NONE,
@@ -1242,17 +1134,13 @@ const FormSelectWithSubview = (props: FormSelectWithSubviewProps & FormikContext
     changeViewButtonLabel,
     subview,
     onChange,
-    tooltipProps,
     ...rest
   } = restProps
 
   const value = get(formik?.values, name)
-  const labelText = !isOptional ? label : `${label} ${optionalLabel}`
   return (
     <FormGroup
-      label={
-        labelText ? <HarnessDocTooltip tooltipId={tooltipProps?.dataTooltipId} labelText={labelText} /> : labelText
-      }
+      label={getFormFieldLabel(label, name, props)}
       labelFor={name}
       helperText={helperText}
       intent={intent}
@@ -1297,7 +1185,7 @@ export interface FormMultiSelectWithSubviewProps extends Omit<IFormGroupProps, '
 }
 
 const FormMultiSelectWithSubview = (props: FormMultiSelectWithSubviewProps & FormikContextProps<any>) => {
-  const { formik, name, isOptional = false, optionalLabel = IsOptionLabel, ...restProps } = props
+  const { formik, name, ...restProps } = props
   const hasError = errorCheck(name, formik)
   const {
     intent = hasError ? Intent.DANGER : Intent.NONE,
@@ -1311,17 +1199,13 @@ const FormMultiSelectWithSubview = (props: FormMultiSelectWithSubviewProps & For
     changeViewButtonLabel,
     subview,
     onChange,
-    tooltipProps,
     ...rest
   } = restProps
 
   const value: MultiSelectOption[] = get(formik?.values, name)
-  const labelText = !isOptional ? label : `${label} ${optionalLabel}`
   return (
     <FormGroup
-      label={
-        labelText ? <HarnessDocTooltip tooltipId={tooltipProps?.dataTooltipId} labelText={labelText} /> : labelText
-      }
+      label={getFormFieldLabel(label, name, props)}
       labelFor={name}
       helperText={helperText}
       intent={intent}