+++ src/components/FormikForm/FormikTooltipContext.tsx
@@ -1,0 +1,7 @@
+import { createContext } from 'react'
+
+interface FormikTooltipContextInterface {
+  formName?: string
+}
+
+export const FormikTooltipContext = createContext<FormikTooltipContextInterface>({})